requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
openpyxl
langchain
langchain-community
langchain-ollama
langchain-core
chromadb
fpdf2
